\name{plot.TSS}
\alias{plot.TSS}
\title{Generic \code{plot} function for TSS Object}
\usage{
\method{plot}{TSS}(x, columns = c("tss", "sensitivity", "specificity"), ...)
}
\arguments{
  \item{x}{object of class \code{TSS}}

  \item{columns}{columns to include in the plot}

  \item{...}{further arguments for \code{plot} function}
}
\description{
Generic \code{plot} function for TSS Object
}

